Convert SLC as deviation factor

+++ sbk-api/src/main/java/io/sbk/logger/impl/Sl4jResultLogger.java
@@ -26,7 +26,7 @@ public class Sl4jResultLogger extends SystemLogger {
 
     private void print(String prefix, double seconds, long bytes, long records, double recsPerSec, double mbPerSec,
                        double avgLatency, long maxLatency, long invalid, long lowerDiscard, long higherDiscard,
-                       int slc1, int slc2, long[] percentileValues) {
+                       long slc1, long slc2, long[] percentileValues) {
         log.info(buildResultString(new StringBuilder(prefix), seconds, bytes, records, recsPerSec, mbPerSec, avgLatency,
                 maxLatency, invalid, lowerDiscard, higherDiscard, slc1, slc2, percentileValues));
     }
@@ -34,7 +34,7 @@ public class Sl4jResultLogger extends SystemLogger {
     @Override
     public void print(double seconds, long bytes, long records, double recsPerSec, double mbPerSec, double avgLatency,
                       long maxLatency, long invalid, long lowerDiscard, long higherDiscard,
-                      int slc1, int slc2, long[] percentileValues) {
+                      long slc1, long slc2, long[] percentileValues) {
         print(prefix, seconds, bytes, records, recsPerSec, mbPerSec, avgLatency, maxLatency,
                 invalid, lowerDiscard, higherDiscard, slc1, slc2, percentileValues);
     }
@@ -42,7 +42,7 @@ public class Sl4jResultLogger extends SystemLogger {
     @Override
     public void printTotal(double seconds, long bytes, long records, double recsPerSec, double mbPerSec,
                            double avgLatency, long maxLatency, long invalid, long lowerDiscard, long higherDiscard,
-                           int slc1, int slc2, long[] percentileValues) {
+                           long slc1, long slc2, long[] percentileValues) {
         print(prefix + "(Total) ", seconds, bytes, records, recsPerSec, mbPerSec, avgLatency, maxLatency,
                 invalid, lowerDiscard, higherDiscard, slc1, slc2, percentileValues);
     }
